Add wrapper to utils
