+++ lib/react_client/js_backed_map.dart
@@ -6,6 +6,7 @@ import 'dart:core';
 import 'dart:js_util' as js_util;
 
 import 'package:js/js.dart';
+import 'package:react/src/react_client/private_utils.dart';
 
 /// A view into a JavaScript object ([jsObject]) that conforms to the Dart [Map] interface.
 ///
@@ -58,12 +59,12 @@ class JsBackedMap extends MapBase<dynamic, dynamic> {
 
   @override
   dynamic operator [](Object key) {
-    return _JsBackedMapValue.unwrapIfNeeded(js_util.getProperty(jsObject, key));
+    return DartValueWrapper.unwrapIfNeeded(js_util.getProperty(jsObject, key));
   }
 
   @override
   void operator []=(dynamic key, dynamic value) {
-    js_util.setProperty(jsObject, key, _JsBackedMapValue.wrapIfNeeded(value));
+    js_util.setProperty(jsObject, key, DartValueWrapper.wrapIfNeeded(value));
   }
 
   @override
@@ -122,28 +123,6 @@ class JsBackedMap extends MapBase<dynamic, dynamic> {
     // hashCode collisions more often, they are completely valid.
     // For more information, see the `Object.hashCode` doc comment.
     return 0;
-  }
-}
-
-/// A wrapper around a value that can't be stored in its raw form
-/// within a JS object (e.g., a Dart function).
-class _JsBackedMapValue {
-  final dynamic value;
-
-  const _JsBackedMapValue(this.value);
-
-  static dynamic wrapIfNeeded(dynamic value) {
-    if (value is Function && !identical(allowInterop(value), value)) {
-      return _JsBackedMapValue(value);
-    }
-    return value;
-  }
-
-  static dynamic unwrapIfNeeded(dynamic value) {
-    if (value is _JsBackedMapValue) {
-      return value.value;
-    }
-    return value;
   }
 }
